+++ sources/record_readers.cpp
@@ -48,7 +48,7 @@ RecordReader::RecordReader(const std::string& file): m_offset(0)
 	// Does not handle compression yet
 }
 
-fsal::Status RecordReader::ReadChecksummed(size_t offset, size_t size, uint8_t* dst)
+fsal::Status RecordReader::ReadChecksummed(uint64_t offset, size_t size, uint8_t* dst)
 {
 	if (size >= SIZE_MAX - sizeof(uint32_t))
 	{
@@ -80,7 +80,7 @@ fsal::Status RecordReader::ReadChecksummed(size_t offset, size_t size, uint8_t* 
 	return true;
 }
 
-fsal::Status RecordReader::ReadRecord(size_t& offset, fsal::MemRefFile* mem_file)
+fsal::Status RecordReader::ReadRecord(uint64_t& offset, fsal::MemRefFile* mem_file)
 {
 	m_file.Seek(offset);
 
@@ -99,7 +99,7 @@ fsal::Status RecordReader::ReadRecord(size_t& offset, fsal::MemRefFile* mem_file
 	return true;
 }
 
-fsal::Status RecordReader::ReadRecord(size_t& offset, std::function<void*(size_t size)> alloc_func)
+fsal::Status RecordReader::ReadRecord(uint64_t& offset, std::function<void*(size_t size)> alloc_func)
 {
 	m_file.Seek(offset);
